+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - `PortScanData.open` property. #3238
+- `{GET,PUT} /api/agent-binaries/<string:os>/masque`. #3249
 
 ### Changed
 - Renamed "Credential collector" to "Credentials collector". #3167